+++ src/internals/souffle/executor/index.ts
@@ -1,8 +1,0 @@
-export {
-  Executor,
-  SyncExecutor,
-  AsyncExecutor,
-  SouffleExecutionResult,
-  SouffleExecutorParams,
-} from "./executor";
-export { ParsedSouffleOutput } from "./results";